+++ src/bfs.c
@@ -8,26 +8,26 @@
 #include "queue.h"
 #include "types.h"
 
-void write_bfs_result(bfs_result r, usize n, FILE *f) {
-  for (usize i = 0; i < n; i += 1) {
-    fprintf(f, "%lu\t%lu\t%lu\n", i, r.distance[i], r.parent[i]);
+void write_bfs_result(bfs_result r, u32 n, FILE *f) {
+  for (u32 i = 0; i < n; i += 1) {
+    fprintf(f, "%u\t%u\t%u\n", i, r.distance[i], r.parent[i]);
   }
 }
 
-bfs_result bfs(csr *adj, usize src) {
+bfs_result bfs(csr *adj, u32 src) {
   // Adjacency matrix should be square matrix
   assert(adj->m == adj->n);
 
   queue *q = queue_new(adj->n);
   bitmap *b = bitmap_new(adj->n);
-  usize *parent = malloc(sizeof(usize) * adj->n);
-  usize *distance = calloc(adj->n, sizeof(usize));
+  u32 *parent = malloc(sizeof(u32) * adj->n);
+  u32 *distance = calloc(adj->n, sizeof(u32));
 
   queue_push(q, src);
   while (!queue_empty(q)) {
-    usize v = queue_pop(q);
-    for (usize j = csr_row_begin(adj, v); j < csr_row_end(adj, v); j += 1) {
-      usize next = adj->c[j];
+    u32 v = queue_pop(q);
+    for (u32 j = csr_row_begin(adj, v); j < csr_row_end(adj, v); j += 1) {
+      u32 next = adj->c[j];
       if (!bitmap_test_set(b, next)) {
         queue_push(q, next);
         parent[next] = v;
