+++ src/main.c
@@ -28,7 +28,7 @@ void test(csr *adj) {
   free(p.parent);
 }
 
-static double bench_one(bfs_func func, csr *adj, usize src) {
+static double bench_one(bfs_func func, csr *adj, u32 src) {
   double start = omp_get_wtime();
   bfs_result r = func(adj, src);
   double end = omp_get_wtime();
@@ -37,9 +37,9 @@ static double bench_one(bfs_func func, csr *adj, usize src) {
   return end - start;
 }
 
-static void bench_multiple(bfs_func func, csr *adj, usize n, char *name) {
+static void bench_multiple(bfs_func func, csr *adj, u32 n, char *name) {
   double sum = 0.0, sum2 = 0.0;
-  for (usize i = 0; i < n; i += 1) {
+  for (u32 i = 0; i < n; i += 1) {
     double dt = bench_one(func, adj, rand() % adj->n);
     sum += dt;
     sum2 += dt * dt;
